fix(vqa): keep the frame loader's result in a type that holds VQAERR_NONE

+++ code/vqalib/loader.cpp
@@ -186,7 +186,9 @@ long VQA_LoadFrame(VQAHandleP *vqap, long flags)
 	int scan_frame;
 	int tocache;
 	VQA_H_FUNC handler;
-	unsigned int rc;
+	// VQAERR_NONE is -1, so a narrower or unsigned holder loses it: it comes back as
+	// 0xFFFFFFFF from a long return and matches no error code the caller knows.
+	long rc;
 	int val4;
 	int fsize;
 	VQA_H_FUNC oldhandler;
@@ -1270,7 +1272,7 @@ long VQA_SeekGroup(VQAHandleP *vqap, long framenum, long groupsize, VQABool prel
 	int bytes_per_frame;
 	int preload_frames;
 	int covered_bytes;
-	unsigned int rc;
+	long rc;
 	bool bool2;
 	int loadflags;
 	long group_end;
